Add new page and new route for view notes
