+++ src/App.js
@@ -11,6 +11,7 @@ import Logout from "./Logout";
 import Aroute from "./Aroute";
 import UserProfile from "./UserProfile";
 import AddNotes from "./AddNotes";
+import ViewNotes from "./ViewNotes";
 
 function App() {
   const HeaderWithNavigation = withNavigation(Header);
@@ -25,6 +26,7 @@ function App() {
             <Route exact path='/login' element={<Login/>}></Route>
             <Route path="/welcome" element={<Aroute><Welcome/></Aroute>} />
             <Route path="/addnotes" element={<Aroute><AddNotes/></Aroute>} />
+            <Route path="/viewnotes" element={<Aroute><ViewNotes/></Aroute>} />
             <Route path="/logout" element={<Aroute><Logout/></Aroute>} />
             <Route path="/profile" element={<Aroute><UserProfile/></Aroute>} />
         </Routes>
